Update patch set 2
Patch Set 2:

> i'm pretty sure you're getting confused by the compat layers that exist in the kernel and don't apply to aarch64 (because it's new enough to not include the legacy 32bit-64bit junk).

to be precise, the compat layer does more than 32bit-64bit management, but that's largely what it's used for, and the esoteric details don't matter to this discussion.  the salient point is that those particular set of compat syscalls don't get used on arm64 and we should not be trying to expose them manually.

Patch-set: 2
Reviewer: Gerrit User 1040868 <1040868@85c56323-6fa9-3386-8a01-6480fb634889>