Update patch set 3
Patch Set 3:

> Patch Set 2: Code-Review-1
> 
> we prefer to not use any new short flags that aren't overtly obvious.  in this case, -A would not pass the "obvious test" imo, so please add a --long-opt instead.

The short options are crowded indeed. Replaced it by --add-group.

> you'll also want to update at least minijail0_cli_unittest.cc to check the new flag.

Hah, I had a look at the test/ subdir and found nothing about testing the cli. Didn't look in the same directory however, so I missed it. Added.

Note that my other pending patch is also adding a long-option. I've made the assumption that it'll be merged eventually, so as the other uses 134 in the switch case, this one uses 135. This can always be fixed later if the other patch never gets merged.

Patch-set: 3