Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> Hi Stéphane,
> 
> Your CL spurred some discussions and we realized that there quite a few use cases that would benefit from passing environments, and not just LD_PRELOAD.
> 
> https://android-review.googlesource.com/c/platform/external/minijail/+/1232493 (landed) and 
> https://android-review.googlesource.com/c/platform/external/minijail/+/1229960 (about to land) add support for passing a new environment to the child. I expect you should be able to rebase on top of those and simplify your code by basically allowing to set LD_PRELOAD in the child using the newly-added mechanisms.
> 
> Thanks!

Thanks for the update!

As the code has now landed, I think I'll refactor this to not be LD_PRELOAD-specific, but allow to clear env and/or pass arbitrary envvars to the child instead (LD_PRELOAD just being another envvar in that case). Probably something around --clear-env / --set-env.

Patch-set: 5
Work-in-progress: false