Update patch set 1
Patch Set 1:

I stumbled upon this issue using minijail on a server with groups that has lots
of members. This is reproducible with a group of ~60 members, as is:

# groupadd test
# for i in $(seq 1 60); do useradd -s /bin/false testuser$i; usermod -a -G test testuser$i; done
# minijail0 -g test /usr/bin/id
Bad group: 'test'

If we use the gid instead, it works:
# minijail0 -g 5608 /usr/bin/id
uid=0(root) gid=5608(test) groups=5608(test)

The logic in lookup_group was to try just once to call getgrnam_r, with a buffer
the size of whatever sysconf() gave us. On my system (a classic Linux x86_64),
it happens to be 1024 only (!). So as this call returns, among other things, the
full list of members, it's easy to understand how the 1024 bytes limit can be
reached easily.

The same problem is reproducible with lookup_user, an easy way to trigger it
is to set an unreasonably long GECOS:

# useradd testuser
# usermod -c $(perl -e "print 'x'x999") testuser
# minijail0 -u testuser /usr/bin/id
Bad user: 'testuser'

Both functions have been modified to use the sysconf() value as a hint for the
first buffer size (as per the manpage), and try with a buffer twice as
big if we get ERANGE until we get it right, or we hit a buffer of 1 MiB in size.

Patch-set: 1