Update patch set 1
Patch Set 1: Code-Review+1

> Patch Set 1:
> 
> Hi, all, 
> 
> The CL is fixing this bug: b/150394301
> 
> Regarding the style, I try not to change, and it did not pass the automatic clang-format test. 

Oh, thanks for pointing this out. We'll change that later so that the scripts provide clang-format-compliant code.

> 
> Crosvm with sandbox on x86 host requires minijail to recognize "statx" in its syscalls table. However, I guess the two *.gen.c under linux-x86 seems to have been generated a while ago by the two gen_*.sh scripts. At that time, likely glibc and/or kernel header did not have statx and bunch of other calls.

Your guess is correct! It's been a while since we've run this. Ideally this file should be re-generated on the CrOS side to avoid depending on what a random developer had on their development machine at that point in time. But refreshing the file sounds like a good idea regardless.

> This CL basically re-ran the two scripts to refresh the two lib*.gen.c under linux-x86,

Patch-set: 1
Reviewer: Gerrit User 1066743 <1066743@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1