Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-2
> 
> These patches already exist in AOSP.

Hi Jorge, 
We are working on merging all patches on android10-qpr2-release to android10-gsi.
We did it when there is a new branch publish to AOSP, such as android10-release and android10-qprN-release.
The patches is empty, means there is no difference between android10-qpr1-release and android10-qpr2-release. However, we still make a merge commit here.
If you still think this is not a good merge. I will ignore the patch on the project.

Patch-set: 1
Reviewer: Gerrit User 1128729 <1128729@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0