libminijail_unittest: Deflake the namespace tests
This change makes `GetProcessSubtreePids` be less flaky: previously it
asserted if it failed to open any of the `/proc/PID/stat` files in
`/proc`. The problem with that is that PIDs can come and go, so that's
inherently racy and makes the test flaky.

Instead, if a particular file cannot be opened with `ENOENT`, treat that
as a normal condition and let it slide.

Bug: None
Test: make 'CXX_BINARY(libminijail_unittest)' && \
      ./libminijail_unittest \
          --gtest_filter=NamespaceTest.test_namespaces \
          --gtest_repeat=100

Change-Id: I3c58b2906d4acddfab27eb6747ca63f79b3347ef