Add instructions for using the seccomp compiler successfully.
There are quite a few requirements to be able to use the seccomp
compiler sucessfully, the trickiest of which is the 'constants.json'
file which is required by the compiler and uses LLVM IR files as input.

Leave enough breadcrumbs so that occasional users can figure out what
they need to do.

Also fix some style nits.

Bug: crbug.com/1162326
Test: generate_constants_json.py -h, compile_seccomp_policy.py -h,
Test: generate_seccomp_policy.py -h
Change-Id: I00ea7d9af3cf854ea8b084048533194a4edabd42