minijail: Don't allow remounting specific dir as more strict
The intended use of the --mount-propagation flag is to allow enforcing
a more strict mount propagation with remount_mode over the filesystem while
allowing for specific mounts to be less strict. This change ensures that
users are doing this and not trying the use the feature the other way
around by making a specific directory more strict than the general
filesystem (which offers little actual security).

Bug: chromium:1102404
Test: make tests
tried to enter minijail with the commandline:
sudo ./minijail0 -v -Kshared --mount-propagation private:/proc /bin/bash
and ensured that it errored with the correct message

Change-Id: Ife992f7a0fed8080afa19499c10a6f688f77e90b