Add overlap handling to redirect_fds().
Previously redirect_fds would clobber mapping source fds if the
destination of a previous mapping collided. This change adds detection
of these collisions and handles it by mapping them ahead of time when
possible or mapping to a temporary otherwise.

It also cleans up some badness in the Rust wrapper that was closing
file descriptors already closed by libminijail, and updates the
fork_remap test to make it easier to debug.

Bug: 1198756
Test: cargo test -- --test-threads=1
Change-Id: I637846dfbe73b73dbb5d218bcfc8464c0cd7d3b4