rust: disable bindgen header with version number
By default, bindgen generates a comment like this:

  /* automatically generated by rust-bindgen 0.56.0 */

If the bindgen version is different than the one used to generate the
checked-in libminijail.rs, the repository will be dirtied even if the
rest of the generated file is the same.

Add the --disable-header-comment argument to bindgen to disable this
header.

Bug: None
Test: cargo build; git diff
Change-Id: I2111461a218cc93223f8dcda7bebb389e5b80a09