tools/compile_seccomp_policy: align BPF buffer
The BPF opcodes emitted by the --output-header-file option are stored in
an array of unsigned char, but if other code accesses the data as an
array of struct sock_filter, the alignment of the buffer may not be
sufficient.  Since struct sock_filter contains a __u32 field, its
natural alignment is 4 bytes, so add an align(4) attribute.

Change-Id: Ic8b2c9d4c7cbdcfc6b69d49c24fa21485fc6a04b
Signed-off-by: Daniel Verkamp <dverkamp@chromium.org>