Tests: Add `source_path` to `test_util.h`
The same code is present in both `syscall_filter_unittest.cc` and
`minijail0_cli_unittest.cc`, and both include `test_util.h`, so it makes
sense to de-duplicate.

This change also removes the empty expression in the use of the ternary
operator, as that is actually a GNU C extension.

Bug: N/A
Test: `make tests`

Signed-off-by: Christian Blichmann <cblichmann@google.com>
Change-Id: Iecfdc49cab629e8aa0374cec9be1219ce66adb9c