Add a flag-protected no-exec mount check on the conf file.
Use syscall statfs() to look up mount flags to check if
the conf file is located in a no-exec mount.

Use a compile time flag to enable or disable the check. Builds that
elects to enable the check can specify a env var through
"export ENFORCE_NOEXEC_CONF=yes".

Fix: 209664870
Bug: 209664870
Test: manual tests and unit tests, "VALGRIND=1 make tests"
Change-Id: I899c8cf7501b64e7c6886951af3672b911ca2378